+++ src/app/item-triage/ItemTriage.tsx
@@ -5,10 +5,11 @@ import ClassIcon from 'app/dim-ui/ClassIcon';
 import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
 import ColorDestinySymbols from 'app/dim-ui/destiny-symbols/ColorDestinySymbols';
 import { ExpandableTextBlock } from 'app/dim-ui/ExpandableTextBlock';
+import { PressTip } from 'app/dim-ui/PressTip';
 import { SetFilterButton } from 'app/dim-ui/SetFilterButton';
 import filterButtonStyles from 'app/dim-ui/SetFilterButton.m.scss';
 import BucketIcon from 'app/dim-ui/svgs/BucketIcon';
-import { t } from 'app/i18next-t';
+import { t, tl } from 'app/i18next-t';
 import { allItemsSelector } from 'app/inventory/selectors';
 import { hideItemPopup } from 'app/item-popup/item-popup';
 import { ItemPopupTab } from 'app/item-popup/ItemPopupBody';
@@ -29,14 +30,19 @@ import { DimItem } from '../inventory/item-types';
 import popupStyles from '../item-popup/ItemDescription.m.scss';
 import styles from './ItemTriage.m.scss';
 import { Factor } from './triage-factors';
-import { getNotableStats, getSimilarItems, getValueColors } from './triage-utils';
+import {
+  getBetterWorseItems,
+  getNotableStats,
+  getSimilarItems,
+  getValueColors,
+} from './triage-utils';
 
 /** whether an item's popup should contain the triage tab */
 export function doShowTriage(item: DimItem) {
   return (
     item.destinyVersion === 2 &&
     (item.bucket.inArmor ||
-      (item.bucket.sort === 'Weapons' &&
+      (item.bucket.sort === 'Weapons' && // there's some reason not to use inWeapons
         item.bucket.hash !== BucketHashes.SeasonalArtifact &&
         item.bucket.hash !== BucketHashes.Subclass))
   );
@@ -78,9 +84,12 @@ export function ItemTriage({ item }: { item: DimItem }) {
     <div className={styles.itemTriagePane}>
       {item.bucket.inWeapons && <WishlistTriageSection item={item} />}
       <LoadoutsTriageSection item={item} />
-      <FactorsTriageSection item={item} />
+      <SimilarItemsTriageSection item={item} />
       {item.bucket.inArmor && item.bucket.hash !== BucketHashes.ClassArmor && (
-        <ArmorStatsTriageSection item={item} />
+        <>
+          <ArmorStatsTriageSection item={item} />
+          <BetterItemsTriageSection item={item} />
+        </>
       )}
     </div>
   );
@@ -164,7 +173,7 @@ function LoadoutsTriageSection({ item }: { item: DimItem }) {
 /**
  * we don't include this section if there's nothing "interesting" to share about this item
  */
-function FactorsTriageSection({ item }: { item: DimItem }) {
+function SimilarItemsTriageSection({ item }: { item: DimItem }) {
   const filterFactory = useSelector(filterFactorySelector);
   const allItems = useSelector(allItemsSelector);
   const itemFactors = getSimilarItems(item, allItems, filterFactory);
@@ -174,11 +183,14 @@ function FactorsTriageSection({ item }: { item: DimItem }) {
     return null;
   }
 
+  // separate section IDs allows separate settings saves
+  const sectionId = `${item.bucket.inArmor ? 'armor' : 'weapon'}-triage-itemcount`;
+
   const fewestSimilar = _.minBy(itemFactors, (f) => f.count)!.count;
   return (
     <CollapsibleTitle
       title={t('Triage.SimilarItems')}
-      sectionId="triage-itemcount"
+      sectionId={sectionId}
       defaultCollapsed={false}
       extra={<span className={styles.factorCollapsedValue}>{fewestSimilar}</span>}
       showExtraOnlyWhenCollapsed
@@ -200,6 +212,94 @@ function FactorsTriageSection({ item }: { item: DimItem }) {
               </span>
             </div>
           ))}
+      </div>
+    </CollapsibleTitle>
+  );
+}
+
+const descriptionBulletPoints = {
+  worse: [tl('Triage.StatWorseArmorDesc'), tl('Triage.PerkWorseArmorDesc')],
+  better: [tl('Triage.StatBetterArmorDesc'), tl('Triage.PerkBetterArmorDesc')],
+} as const;
+
+/**
+ * we don't include this section if there's no strictly better or worse items
+ */
+function BetterItemsTriageSection({ item }: { item: DimItem }) {
+  const filterFactory = useSelector(filterFactorySelector);
+  const allItems = useSelector(allItemsSelector);
+  const betterWorseResults = getBetterWorseItems(item, allItems, filterFactory);
+
+  // turns out exampleItem had no stats... weird?
+  if (!betterWorseResults) {
+    return null;
+  }
+  const { betterItems, artificeBetterItems, worseItems, artificeWorseItems } = betterWorseResults;
+  // nothing interesting = no display
+  if (
+    !betterItems.length &&
+    !artificeBetterItems.length &&
+    !worseItems.length &&
+    !artificeWorseItems.length
+  ) {
+    return null;
+  }
+
+  const rows = [
+    [t('Triage.BetterArmor'), descriptionBulletPoints.better, betterItems],
+    [t('Triage.BetterArtificeArmor'), descriptionBulletPoints.better, artificeBetterItems, true],
+    [t('Triage.WorseArmor'), descriptionBulletPoints.worse, worseItems],
+    [t('Triage.WorseArtificeArmor'), descriptionBulletPoints.worse, artificeWorseItems, true],
+  ] as const;
+
+  return (
+    <CollapsibleTitle
+      title={t('Triage.BetterWorseArmor')}
+      sectionId="better-worse-armor"
+      defaultCollapsed={false}
+      extra={<span className={styles.factorCollapsedValue}>!!</span>}
+      showExtraOnlyWhenCollapsed
+    >
+      <div className={styles.similarItemsTable}>
+        {rows.map(([label, [statDesc, perkDesc], itemCollection, showArtificeDesc]) => {
+          const tooltip = (
+            <>
+              {t('Triage.BetterWorseIncludes')}
+              <ul>
+                <li>
+                  {t(statDesc)}
+                  {showArtificeDesc && (
+                    <span className={styles.artificeExplanation}>
+                      <br />
+                      {t('Triage.AccountsForArtifice')}
+                    </span>
+                  )}
+                </li>
+                <li>{t(perkDesc)}</li>
+              </ul>
+            </>
+          );
+          if (itemCollection.length) {
+            const filter = itemCollection.map((i) => `id:${i.id}`).join(' or ');
+            return (
+              <div className={styles.tableRow} key={label}>
+                <span>
+                  <PressTip tooltip={tooltip} elementType="span" placement="top">
+                    {label}
+                  </PressTip>
+                </span>
+                <span className={styles.count}>{itemCollection.length}</span>
+                <span className={styles.controls}>
+                  <StartCompareButton
+                    filter={`id:${item.id} or ` + filter}
+                    items={itemCollection}
+                    initialItemId={item.id}
+                  />
+                </span>
+              </div>
+            );
+          }
+        })}
       </div>
     </CollapsibleTitle>
   );
@@ -285,10 +385,19 @@ function FactorCombo({
   );
 }
 
-function StartCompareButton({ filter, items }: { filter: string; items: DimItem[] }) {
+function StartCompareButton({
+  filter,
+  items,
+  initialItemId,
+}: {
+  filter: string;
+  items: DimItem[];
+  /** The instance ID of the first item added to compare, so we can highlight it. */
+  initialItemId?: string;
+}) {
   const dispatch = useDispatch();
   const compare = () => {
-    dispatch(compareFilteredItems(filter, items));
+    dispatch(compareFilteredItems(filter, items, initialItemId));
     hideItemPopup();
   };
   const type = items[0]?.typeName;