+++ src/app/utils/item-utils.ts
@@ -1,4 +1,5 @@
 import { factionItemAligns } from 'app/destiny1/d1-factions';
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import {
   D1Item,
   DimItem,
@@ -92,7 +93,7 @@ export const getSpecialtySocketMetadatas = (item?: DimItem): ModSocketMetadata[]
 
 /**
  * combat and legacy slots are boring now. everything has them.
- * this focuses on narrower stuff: raid & nightmare mod
+ * this focuses on narrower stuff: raid & nightmare modslots
  */
 export const getInterestingSocketMetadatas = (item?: DimItem): ModSocketMetadata[] | undefined => {
   const specialtySockets = getSpecialtySocketMetadatas(item)?.filter(
@@ -201,7 +202,7 @@ export function isKillTrackerSocket(socket: DimSocket) {
 }
 
 export interface KillTracker {
-  type: 'pve' | 'pvp';
+  type: 'pve' | 'pvp' | 'gambit';
   count: number;
   trackerDef: PluggableInventoryItemDefinition;
 }
@@ -237,9 +238,12 @@ const d1YearSourceHashes = {
 /**
  * Which "Year" of Destiny did this item come from?
  */
-export function getItemYear(item: DimItem) {
-  if (item.destinyVersion === 2) {
-    const season = getSeason(item);
+export function getItemYear(
+  item: DimItem | DestinyInventoryItemDefinition,
+  defs?: D2ManifestDefinitions
+) {
+  if (('destinyVersion' in item && item.destinyVersion === 2) || 'displayProperties' in item) {
+    const season = getSeason(item, defs);
     return season ? Math.floor(season / 4) + 1 : 0;
   } else if (isD1Item(item)) {
     if (!item.sourceHashes) {
@@ -364,4 +368,13 @@ export function isPlugStatActive(
  */
 export function isD1Item(item: DimItem): item is D1Item {
   return item.destinyVersion === 1;
+}
+
+/** turns an item's list of stats into a dictionary of stats, keyed by stat hash */
+export function getStatValuesByHash(item: DimItem, byWhichValue: 'base' | 'value') {
+  const output: NodeJS.Dict<number> = {};
+  for (const stat of item.stats ?? []) {
+    output[stat.statHash] = stat[byWhichValue];
+  }
+  return output;
 }