+++ docs/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Next
 
+## 7.57.0 <span class="changelog-date">(2023-02-19)</span>
+
+* Add `is:retiredperk` search that highlights items that have a perk which can no longer drop for that item.
 * You can now click a Loadout name in Organizer's Loadouts column to quickly bring up this loadout for editing.
 * When hovering over subclass Aspects in Loadouts and Loadout Optimizer, the tooltip will now show the number of Fragment slots granted.
+* You can now bring up the Armory page for a weapon directly from the search bar by typing a weapon name there and clicking the corresponding entry.
+* Improved the logic for choosing what item to equip when de-equipping an item. DIM will now generally avoid equipping exotics as replacements, and will pay attention to the type of item and your tags.
 
 ## 7.56.0 <span class="changelog-date">(2023-02-12)</span>